Move job server to its own module.
This avoids circular dependencies with the expansion service.
