+++ runners/direct-java/build.gradle
@@ -26,6 +26,7 @@ def dependOnProjects = [":runners:core-construction-java",
                         ":runners:core-java",
                         ":runners:local-java",
                         ":runners:java-fn-execution",
+                        ":runners:java-job-service",
                         ":sdks:java:fn-execution"]
 
 applyJavaNature(
