+++ runners/java-job-service/build.gradle
@@ -16,46 +16,25 @@
  */
 
 plugins { id 'org.apache.beam.module' }
-applyJavaNature(automaticModuleName: 'org.apache.beam.runners.fnexecution')
+applyJavaNature(automaticModuleName: 'org.apache.beam.runners.jobservice')
 
-description = "Apache Beam :: Runners :: Java Fn Execution"
+description = "Apache Beam :: Runners :: Java Job Service"
 
 dependencies {
   compile library.java.vendored_guava_26_0_jre
   compile project(":runners:core-construction-java")
-  provided project(":sdks:java:harness")
   compile project(path: ":model:pipeline", configuration: "shadow")
   compile project(path: ":model:fn-execution", configuration: "shadow")
-  compile project(path: ":sdks:java:core", configuration: "shadow")
+  compile project(":sdks:java:expansion-service")
   compile project(":sdks:java:fn-execution")
   compile project(":runners:core-construction-java")
-  compile project(path: ":vendor:sdks-java-extensions-protobuf", configuration: "shadow")
+  compile project(":runners:java-fn-execution")
   compile library.java.vendored_grpc_1_26_0
-  compile library.java.commons_compress
   compile library.java.slf4j_api
   compile library.java.args4j
-  testCompile project(":sdks:java:harness")
-  testCompile project(":runners:core-construction-java")
-  testCompile project(path: ":runners:core-java", configuration: "testRuntime")
-  testCompile project(path: ":sdks:java:core", configuration: "shadowTest")
   testCompile library.java.junit
   testCompile library.java.hamcrest_core
   testCompile library.java.hamcrest_library
   testCompile library.java.mockito_core
   testRuntimeOnly library.java.slf4j_simple
-}
-
-test {
-  useJUnit {
-    // Exclude tests that need Docker.
-    excludeCategories "org.apache.beam.runners.fnexecution.environment.testing.NeedsDocker"
-  }
-}
-
-task testDocker(type: Test) {
-  group = "Verification"
-  description = "Runs Docker tests"
-  useJUnit {
-    includeCategories "org.apache.beam.runners.fnexecution.environment.testing.NeedsDocker"
-  }
 }
