+++ runners/portability/java/build.gradle
@@ -27,6 +27,7 @@ dependencies {
   compile library.java.vendored_guava_26_0_jre
   compile library.java.hamcrest_library
   compile project(":runners:java-fn-execution")
+  compile project(":runners:java-job-service")
   compile project(path: ":sdks:java:harness", configuration: "shadow")
   compile library.java.vendored_grpc_1_26_0
   compile library.java.slf4j_api
