+++ runners/samza/build.gradle
@@ -42,6 +42,7 @@ dependencies {
   compile project(":runners:core-java")
   compile project(":runners:core-construction-java")
   compile project(":runners:java-fn-execution")
+  compile project(":runners:java-job-service")
   compile library.java.jackson_annotations
   compile library.java.slf4j_api
   compile library.java.joda_time
