+++ settings.gradle
@@ -49,6 +49,7 @@ include ":runners:google-cloud-dataflow-java"
 include ":runners:google-cloud-dataflow-java:examples"
 include ":runners:google-cloud-dataflow-java:examples-streaming"
 include ":runners:java-fn-execution"
+include ":runners:java-job-service"
 include ":runners:jet"
 include ":runners:local-java"
 include ":runners:portability:java"
